bird_augmentation/requirements.txt
streamlit
albumentations
opencv-python-headless
pillow
numpy
